Revert "Implement WsiAnnotations in C++ (#252)"
This reverts commit a497eabe5d206fa063746c6ee236d9206e9f98b6.
