+++ .clang-format
@@ -1,6 +1,4 @@
 ---
-BasedOnStyle: Google
-IndentWidth: 2
 Language: Cpp
 ColumnLimit: 120
 AccessModifierOffset: -4
@@ -71,6 +69,7 @@ IncludeCategories:
 IncludeIsMainRegex: '(Test)?$'
 IndentCaseLabels: false
 IndentPPDirectives: None
+IndentWidth: 4
 IndentWrappedFunctionNames: false
 JavaScriptQuotes: Leave
 JavaScriptWrapImports: true
