+++ dlup/data/transforms.py
@@ -14,7 +14,7 @@ from dlup._exceptions import AnnotationError
 from dlup.annotations import AnnotationClass, AnnotationType
 from dlup.data.dataset import PointType, TileSample, TileSampleWithAnnotationData
 
-_AnnotationsTypes = dlup.annotations.DlupShapelyPoint | dlup.annotations.DlupShapelyPolygon
+_AnnotationsTypes = dlup.annotations.Point | dlup.annotations.Polygon
 
 
 def convert_annotations(
@@ -83,7 +83,7 @@ def convert_annotations(
     has_roi = False
     for curr_annotation in annotations:
         holes_mask = None
-        if isinstance(curr_annotation, dlup.annotations.DlupShapelyPoint):
+        if isinstance(curr_annotation, dlup.annotations.Point):
             coords = tuple(curr_annotation.coords)
             points[curr_annotation.label] += tuple(coords)
             continue
@@ -234,13 +234,13 @@ def rename_labels(annotations: Iterable[_AnnotationsTypes], remap_labels: dict[s
 
         if annotation.annotation_class.annotation_type == AnnotationType.BOX:
             a_cls = AnnotationClass(label=remap_labels[label], annotation_type=AnnotationType.BOX)
-            output_annotations.append(dlup.annotations.DlupShapelyPolygon(annotation, a_cls=a_cls))
+            output_annotations.append(dlup.annotations.Polygon(annotation, a_cls=a_cls))
         elif annotation.annotation_class.annotation_type == AnnotationType.POLYGON:
             a_cls = AnnotationClass(label=remap_labels[label], annotation_type=AnnotationType.POLYGON)
-            output_annotations.append(dlup.annotations.DlupShapelyPolygon(annotation, a_cls=a_cls))
+            output_annotations.append(dlup.annotations.Polygon(annotation, a_cls=a_cls))
         elif annotation.annotation_class.annotation_type == AnnotationType.POINT:
             a_cls = AnnotationClass(label=remap_labels[label], annotation_type=AnnotationType.POINT)
-            output_annotations.append(dlup.annotations.DlupShapelyPoint(annotation, a_cls=a_cls))
+            output_annotations.append(dlup.annotations.Point(annotation, a_cls=a_cls))
         else:
             raise AnnotationError(f"Unsupported annotation type {annotation.annotation_class.annotation_type}")
