--- tests/files/test_different_types_halo.annotations
@@ -1,208 +1,0 @@
-<Annotations>
-  <Annotation Name="Layer 1" Visible="True" LineColor="5634047">
-    <Regions>
-      <Region Type="Rectangle" HasEndcaps="0" NegativeROA="0">
-        <Vertices>
-          <V X="-27900" Y="45644" />
-          <V X="-16121" Y="57423" />
-        </Vertices>
-      </Region>
-      <Region Type="Rectangle" HasEndcaps="0" NegativeROA="0">
-        <Vertices>
-          <V X="-12045" Y="46172" />
-          <V X="8794" Y="49268" />
-        </Vertices>
-      </Region>
-      <Region Type="Ellipse" HasEndcaps="0" NegativeROA="0">
-        <Vertices>
-          <V X="-11592" Y="54402" />
-          <V X="-2154" Y="63840" />
-        </Vertices>
-      </Region>
-      <Region Type="Ellipse" HasEndcaps="0" NegativeROA="0">
-        <Vertices>
-          <V X="1092" Y="55836" />
-          <V X="11814" Y="60216" />
-        </Vertices>
-      </Region>
-      <Region Type="Ruler" HasEndcaps="0" NegativeROA="0">
-        <Vertices>
-          <V X="-28202" Y="63764" />
-          <V X="-14263" Y="63764" />
-        </Vertices>
-      </Region>
-      <Region Type="Ruler" HasEndcaps="0" NegativeROA="0">
-        <Vertices>
-          <V X="-22690" Y="67161" />
-          <V X="-17631" Y="72824" />
-        </Vertices>
-      </Region>
-      <Region Type="Polygon" HasEndcaps="0" NegativeROA="0">
-        <Vertices>
-          <V X="942" Y="67237" />
-          <V X="1093" Y="67237" />
-          <V X="2527" Y="67539" />
-          <V X="2905" Y="67766" />
-          <V X="4339" Y="68294" />
-          <V X="4792" Y="68521" />
-          <V X="5472" Y="68823" />
-          <V X="5774" Y="69049" />
-          <V X="6529" Y="69729" />
-          <V X="6604" Y="69804" />
-          <V X="7586" Y="70937" />
-          <V X="7812" Y="71239" />
-          <V X="8265" Y="72296" />
-          <V X="8341" Y="72447" />
-          <V X="8492" Y="73051" />
-          <V X="8492" Y="73579" />
-          <V X="8416" Y="74032" />
-          <V X="8114" Y="74410" />
-          <V X="7737" Y="74636" />
-          <V X="7208" Y="74712" />
-          <V X="7057" Y="74636" />
-          <V X="6906" Y="74712" />
-          <V X="6529" Y="74787" />
-          <V X="6302" Y="74787" />
-          <V X="5321" Y="74636" />
-          <V X="5019" Y="74636" />
-          <V X="3735" Y="74334" />
-          <V X="3584" Y="74259" />
-          <V X="2603" Y="74108" />
-          <V X="1621" Y="74032" />
-          <V X="413" Y="74032" />
-          <V X="-644" Y="73957" />
-          <V X="-1474" Y="73881" />
-          <V X="-3060" Y="73806" />
-          <V X="-4192" Y="73655" />
-          <V X="-5400" Y="73579" />
-          <V X="-6457" Y="73504" />
-          <V X="-7061" Y="73428" />
-          <V X="-8949" Y="73428" />
-          <V X="-9628" Y="73353" />
-          <V X="-10383" Y="73353" />
-          <V X="-10836" Y="73277" />
-          <V X="-11214" Y="72975" />
-          <V X="-11440" Y="72598" />
-          <V X="-11516" Y="72069" />
-          <V X="-11440" Y="71843" />
-          <V X="-11742" Y="71692" />
-          <V X="-12044" Y="71314" />
-          <V X="-12195" Y="70861" />
-          <V X="-12120" Y="70408" />
-          <V X="-11969" Y="70106" />
-          <V X="-12120" Y="69955" />
-          <V X="-12346" Y="69578" />
-          <V X="-12422" Y="69049" />
-          <V X="-12271" Y="68672" />
-          <V X="-11969" Y="68294" />
-          <V X="-11516" Y="68068" />
-          <V X="-10685" Y="68068" />
-          <V X="-9251" Y="67841" />
-          <V X="-8043" Y="67539" />
-          <V X="-7892" Y="67464" />
-          <V X="-5853" Y="67162" />
-          <V X="-5098" Y="67086" />
-          <V X="-4947" Y="67086" />
-          <V X="-2003" Y="67011" />
-          <V X="-1927" Y="67011" />
-          <V X="942" Y="67237" />
-        </Vertices>
-      </Region>
-      <Region Type="Pin" HasEndcaps="0" NegativeROA="0">
-        <Vertices>
-          <V X="-1399" Y="53194" />
-        </Vertices>
-      </Region>
-      <Region Type="Polygon" HasEndcaps="1" NegativeROA="0">
-        <Vertices>
-          <V X="-27598" Y="81430" />
-          <V X="-27522" Y="81280" />
-          <V X="-27446" Y="81129" />
-          <V X="-27144" Y="80827" />
-          <V X="-26767" Y="80600" />
-          <V X="-25785" Y="80449" />
-          <V X="-24879" Y="80374" />
-          <V X="-24426" Y="80374" />
-          <V X="-23747" Y="80449" />
-          <V X="-22916" Y="80751" />
-          <V X="-22161" Y="81355" />
-          <V X="-21482" Y="82035" />
-          <V X="-20953" Y="82639" />
-          <V X="-20500" Y="83092" />
-          <V X="-20349" Y="83167" />
-          <V X="-19670" Y="83469" />
-          <V X="-18764" Y="83545" />
-          <V X="-17782" Y="83545" />
-          <V X="-16272" Y="83092" />
-          <V X="-15140" Y="82488" />
-          <V X="-14913" Y="82337" />
-          <V X="-13554" Y="81506" />
-          <V X="-11365" Y="80449" />
-          <V X="-9553" Y="79921" />
-          <V X="-7816" Y="79694" />
-          <V X="-7061" Y="79694" />
-          <V X="-5929" Y="80147" />
-          <V X="-4343" Y="81129" />
-          <V X="-2984" Y="82261" />
-          <V X="-1927" Y="83092" />
-          <V X="-1776" Y="83167" />
-          <V X="-870" Y="83696" />
-          <V X="338" Y="83922" />
-          <V X="564" Y="83922" />
-          <V X="1848" Y="83922" />
-          <V X="3962" Y="83696" />
-          <V X="6453" Y="82865" />
-          <V X="8869" Y="81884" />
-          <V X="10530" Y="81053" />
-          <V X="11587" Y="80298" />
-          <V X="12418" Y="79619" />
-          <V X="12569" Y="79468" />
-          <V X="12644" Y="79468" />
-        </Vertices>
-      </Region>
-      <Region Type="Polygon" HasEndcaps="0" NegativeROA="0">
-        <Vertices>
-          <V X="-28474" Y="87177" />
-          <V X="-28245" Y="86948" />
-          <V X="-27557" Y="86604" />
-          <V X="-26526" Y="86604" />
-          <V X="-25610" Y="86490" />
-          <V X="-24121" Y="86604" />
-          <V X="-23663" Y="86604" />
-          <V X="-23548" Y="86719" />
-          <V X="-23319" Y="86948" />
-          <V X="-23090" Y="87292" />
-          <V X="-22861" Y="87635" />
-          <V X="-22288" Y="87979" />
-          <V X="-21142" Y="88552" />
-          <V X="-19997" Y="88781" />
-          <V X="-18508" Y="88781" />
-          <V X="-17248" Y="88781" />
-          <V X="-15530" Y="88093" />
-          <V X="-13926" Y="87177" />
-          <V X="-12666" Y="86375" />
-          <V X="-11749" Y="86032" />
-          <V X="-10375" Y="86146" />
-          <V X="-8427" Y="86375" />
-          <V X="-6824" Y="86719" />
-          <V X="-5449" Y="87063" />
-          <V X="-4304" Y="87406" />
-          <V X="-3158" Y="87750" />
-          <V X="-1211" Y="87979" />
-          <V X="1309" Y="88208" />
-          <V X="3829" Y="88208" />
-          <V X="5089" Y="87979" />
-          <V X="6235" Y="87750" />
-          <V X="7495" Y="87521" />
-          <V X="8984" Y="87063" />
-          <V X="10130" Y="86490" />
-          <V X="11046" Y="85802" />
-          <V X="11619" Y="85344" />
-          <V X="12535" Y="84657" />
-          <V X="12993" Y="84199" />
-          <V X="13108" Y="84084" />
-        </Vertices>
-      </Region>
-    </Regions>
-  </Annotation>
-</Annotations>
