+++ tests/infer_ann.py
@@ -1,0 +1,3 @@
+from dlup.annotations import WsiAnnotations
+
+ann = WsiAnnotations.from_geojson("/Users/jteuwen/annotations.json")
