+++ tox.ini
@@ -1,11 +1,8 @@
 [tox]
-envlist = py311
+envlist = py310, py311
 isolated_build = True
 
 [testenv]
-envdir = {toxworkdir}/env
-setenv =
-    GITHUB_ACTIONS = 1
 deps =
     meson
     meson-python>=0.15.0
@@ -14,10 +11,12 @@ deps =
     spin
     pybind11
     build
-    pyhaloxml
 extras = dev,darwin
 commands =
-    pytest --maxfail=1 --disable-warnings
+    sh -c 'meson setup builddir'
+    sh -c 'meson compile -C builddir'
+    sh -c 'cp builddir/*.so dlup'
+    pytest
 allowlist_externals =
     sh
     pytest
